+++ platypush/pusher/__main__.py
@@ -1,6 +1,0 @@
-from platypush.pusher import main
-
-main()
-
-# vim:sw=4:ts=4:et:
-