+++ CHANGELOG.md
@@ -3,6 +3,12 @@
 All notable changes to this project will be documented in this file.
 Given the high speed of development in the first phase, changes are being reported only starting from v0.20.2.
 
+## [Unreleased]
+
+### Removed
+
+- Removed legacy `pusher` script and `local` backend.
+
 ## [0.20.5] - 2021-03-12
 
 ### Added